+++ package.json
@@ -30,10 +30,11 @@
     "@vercel/ncc": "^0.34.0",
     "bluebird": "^3.7.2",
     "construct-style-sheets-polyfill": "^2.3.5",
+    "kleur": "^4.1.5",
     "mkdirp": "^1.0.4",
     "mocha": "^7.1.1",
     "node-fetch": "^2.6.0",
-    "open": "^7.0.3",
+    "open": "^8.4.0",
     "rimraf": "^3.0.2",
     "rollup": "^2.33.3",
     "rollup-plugin-terser": "^5.3.0",
